// src/app/models/admin/dangkyhoatdongkhoangsan/dangkytralai/dkthamdotralai.model.ts
export class OutputDkThamDoTraLaiModel {
    iddangkythamdo: string;
    dientichthamdo: number;
    dientichtralai: number;
    donvidientich: string;
    lydotralai: string;
    geom: string;
    geovn: string;
    geowgs: string;
    hequychieu: string;
    idhoso: string;
    idgiayphep: string;
    sogiayphep: string;
    ngaycapphep: string;
    loaitralai: number;
    siteid: string;
}

export class InputDkThamDoTraLaiModel {
    iddangkythamdo: string;
    dientichthamdo: number;
    dientichtralai: number;
    donvidientich: string;
    lydotralai: string;
    hequychieu: string;
    idhoso: string;
    idgiayphep: string;
    loaitralai: number;
}